refactor: 관리자, 사용자 view 분리
