+++ src/main/java/roomescape/ui/UserController.java
@@ -1,0 +1,20 @@
+package roomescape.ui;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping
+public class UserController {
+
+    @GetMapping("/reservation")
+    public String getReservationPage() {
+        return "reservation";
+    }
+
+    @GetMapping("/login")
+    public String getLoginPage() {
+        return "login";
+    }
+}
